+++ docs/generate/flags.md
@@ -30,6 +30,7 @@ Those are flags that affect autorest only
 | ----------------------------- | -------------------------------------------------------------------------------------------------------------------------------- |
 | `--output-converted-oai3`     | If enabled and the input-files are `swager 2.0` this will output the resulting OpenAPI3.0 converted files to the `output-folder` |
 | `--skip-semantics-validation` | Disable the semantic validator plugin.                                                                                           |
+| `--memory`                    | Configure the max memory allowed to the autorest core process. Usage: `--memory=8g` or `--memory=8192m` (to set to 8gb)          |
 
 ## Temporary flags
