+++ packages/apps/autorest/jest.config.js
@@ -4,7 +4,7 @@ const defaultConfig = require("../../../jest.default.config");
 
 const config = {
   ...defaultConfig,
-  testMatch: ["<rootDir>/test/**/*.test.ts"],
+  testMatch: ["<rootDir>/src/**/*.test.ts", "<rootDir>/test/**/*.test.ts"],
 };
 
 module.exports = config;
