+++ packages/apps/autorest/package.json
@@ -27,7 +27,7 @@
   },
   "scripts": {
     "start": "node ./dist/src/app.js",
-    "test": "jest --coverage=false",
+    "test": "jest --watch --coverage=false",
     "test:ci": "jest --ci",
     "build": "tsc -p tsconfig.build.json",
     "build:prod": "webpack",
