--- packages/apps/autorest/test/basic-test.test.ts
@@ -1,5 +1,0 @@
-describe("Hello", () => {
-  it("works", () => {
-    expect(1).toEqual(1);
-  });
-});
